+++ fiboa_cli/datasets/template.py
@@ -145,7 +145,7 @@ def convert(output_file, cache = None, **kwargs):
                       Can be used to avoid repetitive downloads from the original data source.
     source_coop_url (str): URL to the (future) Source Cooperative repository. Default: None
     collection (bool): Additionally, store the collection separate from Parquet file. Default: False
-    compression (str): Compression method for the Parquet file. Default: zstd
+    compression (str): Compression method for the Parquet file. Default: brotli
     kwargs: Additional keyword arguments for GeoPanda's read_file() or read_parquet() function.
     """
     convert_(